// front/src/app/Models/movie-request.ts
/*Sign up request model*/
export class Movie {
  movieName: string;
  movieImage: string;
  movieURL: string;
  movieReleaseDate: string;
  movieLength: string;
  directorName: string;

  language: string;
  movieType: string;
  synopsis: string;
  rating: string;
  feedback: number;
  theaterid: string[];
  orders: number;
}